Move unspecific classes out of `comparable` plugin
